+++ DtoCreate/Employees/EmployeeCreate.cs
@@ -9,7 +9,8 @@ namespace MediNet_BE.DtoCreate.Employees
         public DateTime Date_Of_Birth { get; set; }
         public int Gender { get; set; }
         public string PhoneNumber { get; set; }
-        public int SpecialistId { get; set; }
+		public int RoleEmployee { get; set; }
+		public int SpecialistId { get; set; }
         public int ClinicId { get; set; }
     }
 }
