+++ MediNet_BE.csproj
@@ -35,6 +35,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Repositories\Doctors\" />
     <Folder Include="wwwroot\images\clinics\" />
     <Folder Include="wwwroot\images\products\" />
     <Folder Include="wwwroot\images\users\admins\" />
