+++ Models/Employees/Employee.cs
@@ -13,6 +13,7 @@ namespace MediNet_BE.Models.Employees
 		public DateTime Date_Of_Birth { get; set; }
 		public int Gender { get; set; }
 		public string PhoneNumber { get; set; }
+		public int RoleEmployee {  get; set; }
 		public int PositionId { get; set; }
 		public int SpecialistId { get; set; }
 		public int ClinicId { get; set; }
