+++ Program.cs
@@ -8,18 +8,16 @@ using MediNet_BE.DtoCreate.Users;
 using MediNet_BE.Interfaces;
 using MediNet_BE.Interfaces.Categories;
 using MediNet_BE.Interfaces.Clinics;
-using MediNet_BE.Interfaces.Courses;
 using MediNet_BE.Interfaces.Employees;
-using MediNet_BE.Interfaces.Enrollments;
+using MediNet_BE.Interfaces.Employees.Courses;
 using MediNet_BE.Interfaces.Orders;
 using MediNet_BE.Models;
 using MediNet_BE.Models.Employees;
 using MediNet_BE.Models.Users;
 using MediNet_BE.Repositories.Categories;
 using MediNet_BE.Repositories.Clinics;
-using MediNet_BE.Repositories.Courses;
-using MediNet_BE.Repositories.Doctors;
-using MediNet_BE.Repositories.Enrollments;
+using MediNet_BE.Repositories.Employees;
+using MediNet_BE.Repositories.Employees.Courses;
 using MediNet_BE.Repositories.Orders;
 using MediNet_BE.Repositories.Users;
 using MediNet_BE.Services;
